A better lexer
It's now like the total parser combinators, but only recognises and
works at the level of Char. Unlike the parser, we use assertions for
totality, mostly because we're doing some primitive mangling to avoid
allocating Strings. Even so, we can probably do a bit better and remove
some of the assertions eventually...