Add RawImp datatype and s-expression based parser
RawImp is intended as an intermediate form to be used by elaborators for
more expressive languages. It will support raw TT plus implicit
arguments, implicit pattern and type level name bindings, and dotted
patterns.

The idea is that RawImp expressions are checked in some outer context,
and elaborated to typechecked and scope safe Terms.