Improve unbound implicits
Now dealing with unbound implicits which first appear under a binder
in an argument type, e.g. 'b' in '(a -> b) -> a', as long as they don't
depend on that argument type.

Also use a separate namespace for unbound implicits to ensure they don't
clash with any globally defined names.