Add where clauses to PTerm
These elaborate directly to a nested 'let' definition in TTImp.

This made it clear that it's necessary to move options from PTerm to
RawImp.  Since the options affect elaboration in TTImp, it makes more
sense for them to be defined there, and it avoids any guesswork about
what an elaborated name will be (which could vary if it's a local
function or type definition, for example)