Improvements to inference in type constructors
e.g. it's nice to be able to say "Functor : (f : _) => Type" and have
the type of 'f' inferred from the body. If we need to infer a function
type during elaboration, we assume it can only be a simple type, that is
we don't abstract the hole over the whole environment.

(In fact, it would probably be better to abstract the hole for the
function argument and return types over the environment at the point the
hole for f's type was introduced. Perhaps we'll do this later.)