Unbound implicits now default to 0 multiplicity
The 0 multiplicity turns out to be incompatible with forcing. This is
because a forced argument to a constructor needs to be available from
elsewhere, so we either forbid it from being a 0 multiplicity, or take
the simpler solution of just erasing all the 0 multiplicity things when
we're done.

We don't actually do this erasure yet though. It'll feed into the code
generator, certainly.

The pattern match compiler knows about 0 multiplicities, and treats
patterns in 0 positions as variable patterns.