Simpler rules for unbound implicits
Bind them all at the top level, left to right in the order they appear.
This gives unification less work to do, and is easier to explain.
IBindVar still used for binding pattern variables (and we still infer
any additional names if the implicitly bound things have a dependent or
polymorphic type)