Better unification support for interfaces
We need the 'invertibility' information - i.e. that a parameter to a
type class can be inverted to get its arguments - otherwise we won't
even be able to work out the parameter to an interface let alone resolve
the implementation.

Idris 1 enforced this by requiring arguments to interfaces to be
injective. This requirement is not yet implemented, and in any case I
think we can relax it a little bit.