Abstract over full environment in nested blocks
When abstracting over new holes, we skip lets because that breaks
pattern unification (generalising over a thing that is already fixed
means that we'll miss things). However, we do need the name to be in
scope in nested blocks such as case expressions and local function
defintions.

One note: It might be better if 'let' was internal only, and in the high
level notation it got elaborated as a global definition, because that
would simplify a lot of reasoning.