Change elaboration of 'case' block
Instead of trying to work out which names to abstract over, use @
patterns then in the elaborated function we have access to the original
name, and the elaborator can refine it if necessary based on any
dependencies.

This has required a bit of work tidying up as patterns (they now don't
have to be UN names, and there's some processing of terms just to tidy
up the display after elaboration).