Better handling of case/hole/linearity combination
If there's a hole, we don't know if linear variables are going to be
used in the definition of the hole, and we need to know when building
the case type. So, for the moment, we'll guess based on whether the name
appears in the case block. This isn't as accurate as it should be (if
the name is used in a 0 multiplicity, it'll be wrong, and if it's used
but as an implicit it'll be wrong) but for the moment it is at least
sound, and works in most reasonable situations.

It would be better if linearity checking knew to look inside case blocks
when checking and updating hole types, but case blocks aren't part of
the core. We should revisit this.