Refine rules for multiplicities in unification
If a HOF expects a function with argument at multiplicity W, it's okay
to send it a function with argument at multiplicity 1 (because it's more
specific). However, if it expects a function with argument at
multiplicity 0, it can only accept a function with multiplicity 0
because that argument will be erased when compiling.
This patch revises the conversion/unification rules accordingly.