Restrict Int types to 63 bits in chez backend
This is for consistency with a (currently hypothetical) C backend, and C
foreign calls, and because Int is intended to be within a limited range.
We can't use fx+ etc because they throw exceptions on overflow.
This doesn't appear to hurt performance (actually seems to improve it
since we don't need bignums so often...)