More prelude, and elaboration of integers
Since we don't do interface search on the LHS, need to match on integer
literals rather than fromInteger there. This is the same behaviour as in
Idris, but it would be nice if we could do better later.

There's no defaulting mechanism at the moment, so you need to be
explicit about the numeric type you want. This will get annoying fast so
I'll deal with it shortly...