Add a simple defaulting mechanism
A %defaulthint is a hint which only gets run if all other searches fail,
right at the end of elaboration. This allows fromInteger to default to
Integer, in particular.

Integer literals now elaborate as 'fromInteger' on the RHS, but since
'fromInteger' doesn't exist when there's no prelude, it uses either Int
or Integer literals when elaborating with --no-prelude