A bit finer control of auto search
Added "indirect hints", which are hints which might lead to a solution
by solving their arguments. We use these if none of the direct hints
worked, and as with the others they need to be unambiguous.

These are specifically added for chasing parent dictionaries, while
reporting an ambiguity error if there are multiple solutions