Names in types are now in scope in rhs
This works by implicitly adding an @-pattern for the names, as long as
the name isn't overridden by something else.
This is most likely to be useful when writing where clauses which rely
on names in the top level type - there's no longer any need to think
about which names in the type will be in scope due to being parameters,
and which won't, because they'll all be in scope.