Add '?' to mean "infer this value"
This is distinct from "_" in types, which means "I don't care what this
is". The distinction is in what happens when inference fails.
If inference fails for "_", we implicitly bind a new name (just like
pattern matching on the lhs - i.e. it means match anything). If
inference fails for "?", we leave it as a hole and try to fill it in
later. As a result, we can say:

foo : Vect Int ?
foo = [1,2,3,4]

...and the ? will be inferred to be 4. But if we say

foo : Vect Int _
foo = [1,2,3,4]

...we'll get an error, because the '_' has been pi-boound.

So the meaning of "_" is now consistent on the lhs and in types (i.e. it
means infer a value and bind a variable on failure to infer anything).

On the rhs, ? and _ have the same effect, because we don't implicitly
bind anythingi there.