Move contrib code here
This has two purposes: firstly, to allow tinkering with the parser for
performance reasons, in ways we might not want in the Idris libraries
(I've just done this to remove the accessibility proof, which it turns
out hurts performane because the erasure analysis doesn't always manage
to remove higher order functions). Secondly, the path to self hosting is
easier if this repo is as self contained as possible. Once we're there,
and ideally once we have some nice package management, we can tidy up
again!

(This parser modification speeds up processing the prelude by about 2x!)