Reduce environment for elaboration of case blocks
Don't abstract over things in the "outer" environment for pattern
matching purposes (e.g. if we're in a where block, only abstract over
the things which are additionally defined locally).

Case blocks still have access to the whole environment, but we're only
going to match on things in the local environment.