Rename handle/try etc
The shortest names should be reserved for the things which save most
state for backtracking. (Ideally, they'd save all the state that the
caller happens to have, but the IORef based system can't manage this.
One of the prices we pay for doing it the fast way is that we need to
take more care...)