Change delayed elaborators
Traverse the term after elaboration to find delayed elaborators. When we
run the delayed elaborators, it's important to know usage information
about Rig1-bound variables (they could be used anywhere in the term, and
we need the right number for the types of case expressions) and this way
we can collect accurate usage information.