Add Racket to the collection of scheme CGs
Given that it just involves a different header and footer, and a
different command to compile, we might as well... there may be a real
advantage of generating reasonably fast executables and the ability to
use libraries from all the different schemes too.

(There's no test for this yet. Ideally, the test suite should know which
backends are available and only run the relevant tests.)