More betterer resolving of unbound implicits
When we encounter implicits in types (i.e. _) we might solve them, and
if we're going to do so, we need to do that in the biggest scope
possible. Whereas, when we have unbound implicits, we always bind them
in the outermost scope for predictability among other things.

This causes a bit of a conflict, because we don't know in advance
whether a _ is going to be solved or need binding. So, now we make a
note of the holes we introduce this way, and if they are left unresolved
after elaborating a type or a LHS, only then do we mark them as unbound
implicits.