Add further checking in LinearCheck
Up to now this has only checked variable usage correctness. However, it
is also useful if it checks everything else again, because even though
these shouild be correct if unification is correct, it's nice to have an
extra layer with a much simpler (and therefore much more believable!)
type checker.

So, this can catch bugs in the unifier, as well as providing more
confidence in correctness of type checked terms.