Add Core.GetType for eta rule
We were getting the type for eta expansion from the other lambda in
unification, which is wrong because this might not unify with the
application. In this case, we really need to know the types of the
expressions we're unifying, but since it's fiddly to propagate these
(and not useful all that often) we recalculate it when it is needed.