Optimise some run-time identity functions
This looks a bit scary, replacing with believe_me, but since the indices
are erased and there's no other manipulation, it's safe! We will need a
more principled way to do this (e.g. a flag that a function must compile
as identity, and a check that it really does)