Unelab recognises case blocks
This uses the information about pattern matching definitions to spot
when a case block function is called, and converts it back to an ICase.
Needs some care - the interactive case split machinery needs the
desugared version, so also we have a flag to unelab which states whether
to resugar high level structures.