Change recording of usage in holes
We shouldn't assume that the first hole uses a linear binding. Better to
think of each hole as an independent thing, so that when we display hole
types we get the usage count on the assumption the name isn't used in
any other hole. Then once a hole is filled in, we'll see refined usage
information.

This isn't strictly right according to QTT's rules, but on the other
hand it is more useful from a programmer's point of view, which is
really more important in practice.