Move records to TTImp rather than high level
While not *strictly* necessary, this gives us much more scope for the
way to define record updates. In particular, having record update as
part of the TTImp term elaborator means that we'll be able to use update
syntax for dependent types (by generating special purpose case
expressions which know which constructor to match on).

Also, it's a handy syntax to offer to high level languages in general.