Fix for as patterns with linear arguments
We can only use one of the pattern or the as-bound argument, so choose
the pattern. Essentially it's equivalent to a case clause on the
as-bound argument.

This also includes a fix for case expressions which match on a variable
in scope, where that variable should still be accessible, so is as-bound
in the elaborated cases.