Implement 'mutual' blocks
These aren't strictly "mutual" in that the order still matters inside
the block (there are potentially lots of complications if there are
dependencies between types in the block after all). Rather, a mutual
block is elaborated in two passes:

1 Types on the first pass
    i.e. type constructors of data declarations
         function types
         interfaces (in full, since it includes function types)
         records (in full, similarly)
         implementation headers (i.e. note their existence, but not the bodies)
2 Everything else on the second pass

This is a slight difference from Idris 1 (which doesn't do interfaces and
records in full on the first pass) but I think it is more easily
explainable given that it doesn't require knowing how interfaces and
records are elaborated.

If more control is needed, you can forward declare types and functins
(as has always been the case).