Make an effort at improving parse errors
This involves a couple of new failure combinators, 'fatalError' and
'mustWork'. 'fatalError' allows us to abort the whole parse if we detect
something that can't possibly be valid. Similarly, 'mustWork' abandons
the whole parse if its argument has an error.

This gives us a bit more control over the errors which get propagated
back to the user, and in particular it allows us to abort immediately if
blocks end prematurely or specific keywords or symbols are missing.
'fail' is now only used for reporting that the current alternative has
failed.