Initial implementation of 'with'
Still lots to do but the basic mechanism is there. TODO: "Magic" with,
abstracting over the value being matched in the generated function type;
recursive notation on rhs; allowing empty clauses on the lhs if the same
as the parent clause; improving error messages for clause match
failures; refactoring of the clause elaborator.