Experiment: borrowing linear things
This is an experiment, and a work in progress which I may return to
later, so I've put it behind a '%language Borrowing' extension. It adds
a new multiplicity annotation '&' indicating that the argument is
'borrowed' - i.e. you can inspect it, but not consume it.  Functions
which borrow linear names must be fully applied and return a concrete
type, to ensure that the borrowed value doesn't outlive the original
value.

This is to allow read only references to linear things (e.g. lookup in
arrays or reading from file handles) in a nicer way that having to
thread the reference. There's still a few things to resolve (fork and
Lazy cause small complications, borrowed arguments in data constructors
don't really make sense, and records could get more interesting).

There is a translation from QTT with borrowing to pure QTT, by
transforming to A-Normal Form, which is what convinces me this is a
reasonable thing to do. It does make the typing rules rather less
pretty, however.