Coverage checking for empty definitions
A clause is impossible if a pattern variable has an empty type, so look
out for these when checking missing cases. A type is empty if it has no
constructors, or if all of the available constructors conflict with the
type (that is with a different constructor at the head of an index).