Inline dictionaries for totality check
This involves a flag 'TCInline' which means that a definition should be
inlined before calculating size change, but not inlined otherwise. It's
potentially dangerous because it might mean something is inlined away
that would otherwise be non-terminating if not used with care, so it
should only be used in cases where it won't change termination
behaviour.

It might be tempting to *always* inline dictionaries, but this risks
making definitions bigger in some cases so I'm not yet certain it's a
good idea. Probably better to implement dictionary specialisation via
partial evaluation as in Idris 1.