Strict positivity condition
To do this, we need to keep track of which data types are declared
mutually, so when one is forward declared note this in the context.
A type T is considered mutual with any other type which either wasn't
defined at the point of T's declaration, or at the point of T's
definition.