More tail recursion in evaluator
This means things can evaluate for longer without crashing! Also
introduce a 'boredom threshold' or 'fuel' for the evaluator to cut off
when evaluation has taken a while during type checking. This will
usually indicate an error but should be made configurable.

(The type checker doesn't yet pay attention to totality so will happily
evaluate non-total things. There's plenty of situations where this will
actually be fine, and it's even useful if we want to reason about
partial things, but it's at the expense of decidable type checking, so a
Decision Will Need To Be Made here later...)