Some unification fixes/improvements
The optimisation for finding argument types to help with disambiguation
was sometimes wrong in the case of dependent functions - in this case
the result type mustn't depend on any of the arguments or unification
might resolve things it shouldn't.
Also fixed an issue with the pattern unification rule, and improved
inference for lambdas where the result type is known.