Fixed inconsistency between REPL and -x for casting strings to nums
string->number in the scheme backends also parse strings that include
prefixes. Scheme lets you use prefixes to specify the radix and
whether or not the number is exact. The REPL on the other hand fails
to parse these prefixes. To fix this, strings with prefixes in them
are now destroyed. Once destroyed they will be unable to parse into a
number.