Set the 'Overloadable' flag on interface methods
I'm not especially keen on this, but so much Idris is written with this
assumption that I think it might be necessary! The effect is: if a name
could be either an interface method, or a function 'f' returning
something in a type T, and we're in a context where the expected type is
a T, then always prefer 'f'.

This allows ad-hoc overloading of names like (>>=), (<+>), (<|>) to work
just like Idris 1, which we do a lot when writing specialised versions
for types which are more dependent than the Prelude interfaces allow.

We should revisit this later (possibly after a redesign of the Prelude,
or once it's easier to drop in alternative preludes).