Tweak the way we deal with Rig0 in case trees
At compile time, we might want to match on Rig0 things, but only to pull
out pattern variables which are also Rig0s. This will allow slightly
more freedom in what we're allowed to say is erased, in particular it'll
allow 'collapsible' things like equality proofs and accessibility
predicates to become erasable.