Allow matching on Rig0, at compile time only
In the run time case tree, if we find we need to split on a Rig0 value,
report an error since it won't be there at run time. Allowing matching
at compile time gives us some more flexibility in what is allowed to be
marked 0 - as long as we don't need to case split at run time, we
already know any constructor fields will be erased (because they are 0
multiplicity) so we can flag proofs (notably Refl!) as erasable.