Support matching on function types
We can only match directly on the input type, because the return type
depends on a value of the input type which we don't have when building
the case tree. It is still possible to match indirectly (e.g. via a case
block, say) as long as you can come up with a value of the input type.