Teach linearity checking about case blocks
This is quite fiddly, but helpful, because we no longer need to predict
in advance what the variable usage of a case block is going to be.
Rather, we can find out afterwards. So if there's two case blocks, for
example, a variable can be used in one or the other, without needing to
arbitrarily pick one.

This also means we can spot usages for things that are implicitly
resolved, rather than needing them to be explicitly used.

A disadvantage is that 'case' types are no longer completely accurate,
but since that's only about the linearity (now checked elsewhere) and
case blocks get inlined for compilation anyway, I don't think it's a
huge issue.