Auto search only resolves locals at RigW
Since searches can be resolved at any arbitrary point during
elaboration, there's no guarantee that we'll put out the right thing if
we allow searches to be resolved by linear variables. Also, we don't
have complete linearity information at this point anyway (in case
blocks, in particular, it's only when we do the final linearity check
that we knoww whether it's okay for a variable to be used or not) so it
could pull out the wrong answer.

So, better all around to disallow it, at least for the moment - auto is
intended for proofs and interfaces in any case.