Make 'case' work under unbound implicits
This is an annoying problem, because to elaborate case expressions we
need to build a type for the case function using the current
environment. But, if there are unbound implicits, we don't have the
complete current environment yet (it gets resolved at the end of
elaboration).

It's reasonably easy to calculate which implicits are needed, but hard
to set up a new term in the right environment. So, given the choice
between rearranging the environment appropriately (fiddly because a lot
of things depend on the environment, and hard to understand), completely
changing how case elaboration works (please not right now...) and
throwing an exception saying which implicits were needed and retrying,
I've picked the last one. It's a bit ugly, but by far the easiest to see
what's going on.

One decision implicitly taken by this commit: unbound implicits which
are introduced in a case branch scope only over that case branch.